split StateLessFunctions.tsx
